Using unix paths when needed
