+++ conans/test/util/tools_test.py
@@ -29,9 +29,10 @@ class RunnerMock(object):
         self.command_called = None
         self.return_ok = return_ok
 
-    def __call__(self, command, output, win_bash=False):  # @UnusedVariable
+    def __call__(self, command, output, win_bash=False, subsystem=None):  # @UnusedVariable
         self.command_called = command
         self.win_bash = win_bash
+        self.subsystem = subsystem
         return 0 if self.return_ok else 1
